- Updated navigation system to nav type safety

+++ shared/src/commonMain/kotlin/presentation/ui/main/categories/CategoriesScreen.kt
@@ -35,7 +35,7 @@ fun CategoriesScreen(
     state: CategoriesState,
     events: (CategoriesEvent) -> Unit,
     popup: () -> Unit,
-    navigateToSearch: (Int) -> Unit,
+    navigateToSearch: (Long) -> Unit,
 ) {
 
     DefaultScreenUI(
